Save users last activity
Every ten seconds we ping the server to set the last activity of a user
to track if they're using Spectrum right now or not.

This will be useful for a variety of things, right now I need it for
email notifications and we can also use this to online status, grab users if they don't
come back etc.