Implement Apollo Engine
Engine supports query batching since the last release (yay!), so I
figured we might as well implement it now.

Note that this does not implement any caching yet, it just enables
Engine so we can get the nice performance overviews again that we had
with Optics. I'll do the caching stuff in another PR.