Quoted message replies
- [x] Add new field to `Message` in GraphQL called `quotedMessage`, add
resolver for it
- [ ] Build UI/UX for it

See #2142, this seems like a relatively quick feature to build that adds
a ton of value for long conversations, and helps them become more
manageable. I picked this up because somebody literally posted a
screenshot of a message that was further up in a thread to reply to it,
so it seems like this would be a useful feature. (also, apart from the
Snapchat clone WhatsApp has yet to launch a single feature that didn't
make 100% sense, those folks over there are :fire: with product
decisions)