Create CODE_OF_CONDUCT.md
Congrats on being open source guys! Wanted to just add a template Code of Conduct for all your new contributors adapting it from contributor-covenant.org. I decided to use the email hi@spectrum.chat. Give it a look, but hopefully being as general as it is, it should work for now.