Create android-friendly wrapper around draft-js-plugins-editor
Instead of making the Android fix specific to the chat input this patch
moves the Android fix a level deeper inside of a custom wrapper
component around draft-js-plugins-editor.

This custom wrapper component conditionally renders either the standard
draft-js-plugins-editor, or a vanilla text input if on Android. This
should theoretically allow us to reuse the same fix for the main thread
composer so folks can also create threads on Android.