[RFC] React Primitives
This commit introduces [react-primitives](https://github.com/lelandrichardson/react-primitives) to the mobile app codebase.

react-primitives exposes a limited set of ReactNative-alike APIs
(among them `<View>`, `<Text>` and `<Touchable>`) and makes them
work across many React render targets.

This commit adds the react-primitives dependency and reworks the
mobile `<Text>` component to be implemented with primitives. **This allows
the `<Text>` component to be rendered in the native app, in the web app,
and even in Sketch and ReactVR—it's now truly universal!**

As you can see in the changes, styled-components also has built-in
support for `react-primitives` by importing from
`styled-components/primitives` instead of `/native`.

While there's always parts of the apps that we'll have to build with
their native APIs, we should be able to build a large portion of our
mobile component library based on primitives, allowing us to potentially
reuse a lot of the work across our platforms in the future.