[PoC] Better thread scrolling on web
This is a first proof of concept for better thread scrolling on the web.
Instead of using a ton of custom code to scroll to the bottom of threads
at the right time, we completely invert the DOM node with a `transform`.
This makes the browser automatically stick to the bottom of the scroll,
no matter if the images load in slowly afterwards or not.

This is not perfect, only a proof of concept, with lots of bugs to be
fixed.