Fix db load on usersCommunities table
Every time anybody loaded the community view we would run a query that
would go through all the members of a community (that was quick), then
filter them by e.g. `isMember = true` (that was slow for big communities)
and then order them by `r.desc('reputation')`. (that was also slow)

Together, the filter and the orderBy meant that we would read a couple
tens of thousands of records when somebody opened a large communities
community view. Multiply that by concurrent users pointed at the
platform and you've got a perfect storm of traffic we can't handle.

This patch fixes it by introducing a couple new indexes on the
usersCommunities table:

- `communityIdAndIsMember`
- `communityIdAndIsModerator`
- `...IsOwner`
- `...IsTeamMember` (moderator or owner)

Those four indexes allow us to efficiently query all the data we
commonly need.