Stop fetching and deprecate Thread.participants
This is a really expensive fields to load from the db, we load it for
every thread all the time but really don't use it anymore since we
removed the face piles.

This patch removes all stray usage of Thread.participants from the
frontend (with 0 functionality going missing), stops fetching it from
the API and marks the field as deprecated.